+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [freemarker] added load-bundled-functions property
 - [freemarker] Changed latest freemarker constant to 2.3.33
+- [playground-quarkus] quarkus-version set to 3.13.2
 
 ## [8.5.1] - 2024-08-09