+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [8.5.2] - 2024-08-14
+
 ### Changed
 
 - [freemarker] added load-bundled-functions property